requirements.txt
beautifulsoup4
Django
django-cron
ipython
lxml
mysqlclient
pylibmc
PyYAML
fake-useragent
PyExecJS
